adjust code to pass tests
